Modularize server into multiple files
