+++ Dockerfile
@@ -16,4 +16,4 @@ EXPOSE 3000
 HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
   CMD wget --quiet --tries=1 --spider http://localhost:3000 || exit 1
 
-ENTRYPOINT ["node", "index.js"]
+ENTRYPOINT ["npm", "start"]
